HU-113-fix: bugs for presentation (#63)
* fix: bugs for presentation


* fix: bugs for presentation

* fix: testing

---------

+++ backend/src/main/java/com/easyschedule/backend/academico/oferta_materia/dto/OfertaDetalleResponse.java
@@ -5,7 +5,7 @@ import java.util.List;
 public record OfertaDetalleResponse(
     Long mallaMateriaId,
     String nombreMateria,
-    int creditos,
+    Short creditos,
     List<String> prerequisitos,
     List<OfertaMateriaResponse> gruposDisponibles
 ) {}
